+++ platform/java/jni/pdfdocument.c
@@ -1542,11 +1542,11 @@ FUN(PDFDocument_addEmbeddedFile)(JNIEnv *env, jobject self, jstring jfilename, j
 JNIEXPORT jstring JNICALL
 FUN(PDFDocument_getEmbeddedFileParams)(JNIEnv *env, jobject self, jobject jfs)
 {
-	return FUN(PDFDocument_getFilespecParams)(env, self, jfs);
+	return FUN(PDFDocument_getFileSpecParams)(env, self, jfs);
 }
 
 JNIEXPORT jstring JNICALL
-FUN(PDFDocument_getFilespecParams)(JNIEnv *env, jobject self, jobject jfs)
+FUN(PDFDocument_getFileSpecParams)(JNIEnv *env, jobject self, jobject jfs)
 {
 	fz_context *ctx = get_context(env);
 	pdf_obj *fs = from_PDFObject_safe(env, jfs);