+++ platform/java/src/com/artifex/mupdf/fitz/PDFDocument.java
@@ -246,7 +246,7 @@ public class PDFDocument extends Document
 
 	public native PDFObject addEmbeddedFile(String filename, String mimetype, Buffer contents, long created, long modified, boolean addChecksum);
 	public native PDFEmbeddedFileParams getEmbeddedFileParams(PDFObject fs);
-	public native PDFFilespecParams getFilespecParams(PDFObject fs);
+	public native PDFFileSpecParams getFileSpecParams(PDFObject fs);
 	public native Buffer loadEmbeddedFileContents(PDFObject fs);
 	public native boolean verifyEmbeddedFileChecksum(PDFObject fs);
 	public native boolean isEmbeddedFile(PDFObject fs);
@@ -259,14 +259,14 @@ public class PDFDocument extends Document
 		return addEmbeddedFile(filename, mimetype, contents, createdTime, modifiedTime, addChecksum);
 	}
 
-	public static class PDFFilespecParams {
+	public static class PDFFileSpecParams {
 		public final String filename;
 		public final String mimetype;
 		public final int size;
 		public final Date creationDate;
 		public final Date modificationDate;
 
-		protected PDFFilespecParams(String filename, String mimetype, int size, long created, long modified) {
+		protected PDFFileSpecParams(String filename, String mimetype, int size, long created, long modified) {
 			this.filename = filename;
 			this.mimetype = mimetype;
 			this.size = size;
@@ -275,7 +275,7 @@ public class PDFDocument extends Document
 		}
 	}
 
-	public static class PDFEmbeddedFileParams extends PDFFilespecParams {
+	public static class PDFEmbeddedFileParams extends PDFFileSpecParams {
 		protected PDFEmbeddedFileParams(String filename, String mimetype, int size, long created, long modified) {
 			super(filename, mimetype, size, created, modified);
 		}