+++ docs/src/changes.rst
@@ -37,6 +37,12 @@ To minimise the impact of such changes, we undertake to list the :title:`API` ch
 Changes from 1.25 to 1.26
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 - `fz_skew_detect` renamed `fz_detect_skew`.
+- `Filespec` renamed `FileSpec` to be consistent across bindings and documentation:
+    - `PDFAnnotation.hasFilespec` renamed `PDFAnnotation.hasFileSpec`
+    - `PDFAnnotation.getFilespec` renamed `PDFAnnotation.getFileSpec`
+    - `PDFAnnotation.setFilespec` renamed `PDFAnnotation.setFileSpec`
+    - `PDFDocument.getFilespecParams` renamed `PDFDocument.getFileSpecParams`.
+    - `PDFObject.isFilespec` renamed `PDFObject.isFileSpec`.
 
 
 Changes from 1.24 to 1.25